+++ packages/types/src/localization.ts
@@ -102,6 +102,7 @@ type _LocalizationResource = {
       title: LocalizationValue;
       titleCombined: LocalizationValue;
       subtitle: LocalizationValue;
+      subtitleCombined: LocalizationValue;
       actionText: LocalizationValue;
       actionLink: LocalizationValue;
       actionLink__use_phone: LocalizationValue;
@@ -175,6 +176,7 @@ type _LocalizationResource = {
       title: LocalizationValue;
       titleCombined: LocalizationValue;
       subtitle: LocalizationValue;
+      subtitleCombined: LocalizationValue;
       actionText: LocalizationValue;
       actionLink: LocalizationValue;
       actionLink__use_email: LocalizationValue;
@@ -757,6 +759,7 @@ type _LocalizationResource = {
     membersPage: {
       detailsTitle__emptyRow: LocalizationValue;
       action__invite: LocalizationValue;
+      action__search: LocalizationValue;
       start: {
         headerTitle__members: LocalizationValue;
         headerTitle__invitations: LocalizationValue;